fix: fixed bug from splitting a node when it has too many keys
Previously, when adding a new key and the node was almost full, the function didn't sort all the keys, just the fist n (= degree) ones. As a result => the middle key wasn't the correct one. That's why I changed the capacity of the array, to can safely add a new node and sort all the keys before splitting.